Moves: Switch to Jetpack menu card logic to JetpackFeatureCardHelper
Moves: the switch to jetpack menu card logic to
JetpackFeatureCardHelper which seems much more relatable

Removes: un used dependencies in test class
